Fixed a parameter incompatibility showing only on some systems. (why?)
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@19262 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/sid/resid-fp.cc
@@ -75,7 +75,7 @@ static int residfp_init(sound_t *psid, int speed, int cycles_per_sec)
     sampling_method method;
     char model_text[100];
     char method_text[100];
-    double passband;
+    float  passband;
     int filters_enabled, model, sampling, passband_percentage;
 
     if (resources_get_int("SidFilters", &filters_enabled) < 0)
